requirements.txt
streamlit==1.36.0
pinecone-client==3.3.0
sentence-transformers==2.7.0
google-generativeai==0.3.2
gitpython==3.1.43
tiktoken==0.7.0
python-dotenv==1.0.1